Pass parameters to make the lottery more a function
The main body of the code is now the function lottery(), which gets the
essential lottery parameters passed to it. A new main() function sets up
these parameters and calls lottery(). This makes it a bit clearer to
manually edit the script to run the lottery with the right parameters.

By making the function parameters, it will also be easier to call the
lottery from an external script, for example a simple GUI.